Reading attribute in PE symbols
