+++ src/Compilers/CSharp/Test/Emit/CodeGen/CodeGenInParametersTests.cs
@@ -681,61 +681,5 @@ class Program
   IL_0010:  ret
 }");
         }
-
-        [Fact]
-        public void RefReadOnlyParameterIsWrittenToMetadata_IfItExistsInSameAssembly()
-        {
-            var text = @"
-namespace System.Runtime.InteropServices
-{
-    public class RefReadOnlyAttribute : System.Attribute { }
-}
-class Test
-{
-    public void M(ref readonly int x) { }
-}
-";
-
-            CompileAndVerify(text, symbolValidator: module =>
-            {
-                var attribute = module.ContainingAssembly
-                    .GetTypeByMetadataName("Test")
-                    .GetMethod("M")
-                    .GetParameters().Single()
-                    .GetAttributes().Single();
-
-                Assert.Equal("RefReadOnlyAttribute", attribute.AttributeClass.MetadataName);
-            });
-        }
-
-        [Fact]
-        public void RefReadOnlyParameterIsWrittenToMetadata_IfItExistsInADifferentAssembly()
-        {
-            var codeA = @"
-namespace System.Runtime.InteropServices
-{
-    public class RefReadOnlyAttribute : System.Attribute { }
-}";
-
-            var referenceA = CreateCompilationWithMscorlib(codeA).VerifyDiagnostics().ToMetadataReference();
-
-            var codeB = @"
-class Test
-{
-    public void M(ref readonly int x) { }
-}
-";
-
-            CompileAndVerify(codeB, additionalRefs: new[] { referenceA }, symbolValidator: module =>
-            {
-                var attribute = module.ContainingAssembly
-                    .GetTypeByMetadataName("Test")
-                    .GetMethod("M")
-                    .GetParameters().Single()
-                    .GetAttributes().Single();
-
-                Assert.Equal("RefReadOnlyAttribute", attribute.AttributeClass.MetadataName);
-            });
-        }
     }
 }
